Two fixes in this update.
Firstly, a check on the number of points in the clock in AnitaEventCalibrator. Broken events were causing data calibration scripts to fall over. Rolling average temperature correction is not updated in the case of crappy event clocks.

Secondly, the constructor in UsefulAnitaEvent that takes a PrettyAnitaHk wasn't setting the fFromCalibratedAnitaEvent flag in UsefulAnitaEvent. This was causing crashes but is fixed now. Happy analysing!