Beginning of CrossCorrelator overhaul. Remove silly FilterBits.
This is the beginning of bringing the CrossCorrelator internals into
line with the AnitaAnalysisFramework method of doing
reconstruction. Expecting all I/O to be FilteredAnitaEvents. I will
try and maintain backwards compatability, but trying to do that in the
past has made things a real mess. This is my library. If it annoys
you, fork it and make your own.