Add QualityCut class, implements and stores my analysis cuts better
This base class and its derivatives (SurfSaturationCut and
SelfTriggeredBlastCut) are to be used prior to any kind of rolling
average power spectra (FourierBuffer), since self triggered
blasts (and surf saturation spikes) contain bucket loads of power
which distort the averages.

For analysis, each instance of the QualityCut contains the relevent
event value, the limits for the cut, a description string. The
function which derived the figures of merit is also defined there.

All classes have the ROOT I/O macros and appear in the LinkDef.
The base class (QualityCut) is required to prevent ROOT complaining
about not knowing what the base class was and outrageously threatening
to not save the associated info.

Additionally this commit constifies the max-min finding TGraph
functions in RootTools.