Add RayleighMonitor filter, RayleighHist, bug fixes in FourierBuffer
RayleighMonitor filter is supposed to monitor the
RayleighDistributions and will be inherited from when I get round to
actually doing something with the data. The functionality I have a
very hacky partial integration with MagicDisplay which is very useful
for checking things vaguely work.

I have added a RayleighHist class which inherits from TH1D. This will
be useful for MagicDisplay and maybe being able to tweak the fitter
for speed under the hood. Currently things are not very
functional. There is no fitting.