Switched RayleighHist to do average over events, rather than time.
This makes things significantly quicker, allowing a single allocation
of memory with a saner structure. Fourier buffer now records the
chiSquare and ndf of the "RayelighHist::Eval()" function. This
function evaluates the rayleigh distribution without the need for any
fitting as the mean and number of events leave no degrees of freedom
for fitting. I'll try and evaluate how well this actually does.

I moved the TPad dividing function into RootTools as it now gets used in
multiple places. This change propogates in an uninteresting way
through most of the different files in this commit.